Move more demos into doc
