+++ docs/examples/basics/config.json
@@ -1,12 +1,16 @@
 {
    "order": [
      "demo_Buniform_Ezero.jl",
+     "demo_electron_proton.jl",
+     "demo_multiple.jl",
      "demo_ExB_drift.jl",
      "demo_gravity_drift.jl",
      "demo_gradient_B.jl",
      "demo_curvature_B.jl",
      "demo_FLR.jl",
      "demo_polarization_drift.jl",
-     "demo_magneticmirror.jl"
+     "demo_magneticmirror.jl",
+     "demo_proton_dipole.jl",
+     "demo_currentsheet.jl"
    ]
 }
