+++ docs/examples/basics/demo_ExB_drift.jl
@@ -1,17 +1,14 @@
 # ---
-# title: Grad-B drift
-# id: demo_gradB
+# title: E×B drift
+# id: demo_ExB
 # date: 2023-04-19
 # author: "[Tiancheng Liu](https://github.com/TCLiuu); [Hongyang Zhou](https://github.com/henry2004y)"
 # julia: 1.9.0
-# description: Simple magnetic field gradient drift demonstration using Makie
+# description: Simple ExB drift demonstration using Makie
 # ---
 
-# This example demonstrates a single proton motion under a non-uniform B field with gradient ∇B ⊥ B.
-# The orbit of guiding center includes some high order terms, it is different from the
-# formula of magnetic field gradient drift of some textbooks which just preserves the first
-# order term.
-# More theoretical details can be found in Introduction to Plasma Physics and Controlled Fusion by F. F. Chen, and Fundamentals of Plasma Physics by Paul Bellan.
+# This example demonstrates a single proton motion under uniform E and B fields.
+# More theoretical details can be found in Introduction to Plasma Physics and Controlled Fusion by F. F. Chen and Computational Plasma Physics, Toshi Tajima.
 
 using JSServe: Page # hide
 Page(exportable=true, offline=true) # hide
@@ -22,36 +19,33 @@ using TestParticleMakie
 using OrdinaryDiffEq
 using StaticArrays
 using LinearAlgebra
-using ForwardDiff: gradient
-import WGLMakie as WM
+using WGLMakie
 
-function grad_B(x)
-    return SA[0, 0, 1e-8+1e-9 *x[2]]
+function uniform_B(x)
+    return SA[0, 0, 1e-8]
 end
 
 function uniform_E(x)
     return SA[1e-9, 0, 0]
 end
 
-abs_B(x) = norm(grad_B(x))
-
 ## trace the orbit of the guiding center
 function trace_gc!(dx, x, p, t)
-    q, m, E, B, sol = p
+    _, _, E, B, sol = p
     xu = sol(t)
-    gradient_B = gradient(abs_B, x)
     Bv = B(x)
     b = normalize(Bv)
     v_par = (xu[4:6]⋅b).*b
-    v_perp = xu[4:6] - v_par
-    dx[1:3] = m*norm(v_perp)^2*(Bv×gradient_B)/(2*q*norm(Bv)^3) + (E(x)×Bv)/norm(Bv)^2+v_par
+    B2 = sum(Bv.^2)
+    dx[1:3] = (E(x)×Bv)/B2 + v_par
 end
 
 x0 = [1.0, 0, 0]
 v0 = [0.0, 1.0, 0.1]
 stateinit = [x0..., v0...]
 tspan = (0, 20)
-param = prepare(uniform_E, grad_B, species=Proton)
+## E×B drift
+param = prepare(uniform_E, uniform_B, species=Proton)
 prob = ODEProblem(trace!, stateinit, tspan, param)
 sol = solve(prob, Tsit5(); save_idxs=[1,2,3,4,5,6])
