+++ docs/examples/basics/demo_electron_proton.jl
@@ -1,13 +1,24 @@
-# Tracing charged particle in a static EM field.
-#
-# Hongyang Zhou, hyzhou@umich.edu
+# ---
+# title: Proton and electron in a static EM field
+# id: demo_proton_electron
+# date: 2023-04-20
+# author: "[Hongyang Zhou](https://github.com/henry2004y); [Tiancheng Liu](https://github.com/TCLiuu)"
+# julia: 1.9.0
+# description: Simple particle trajectories under uniform B and zero E
+# ---
+
+# This example demonstrates tracing a single proton and electron motion under a uniform B field. The E field is assumed to be zero such that there is no particle acceleration.
+
+using JSServe: Page # hide
+Page(exportable=true, offline=true) # hide
 
 using TestParticle
 using Meshes
 using OrdinaryDiffEq
-using Plots
+using TestParticleMakie
+using WGLMakie
 
-## Initialize grid and field
+### Initialize grid and field
 
 x = range(-10, 10, length=15)
 y = range(-10, 10, length=20)
@@ -26,7 +37,7 @@ grid = CartesianGrid((length(x)-1, length(y)-1, length(z)-1),
    (x[1], y[1], z[1]),
    (Δx, Δy, Δz))
 
-## Initialize particles
+### Initialize particles
 
 x0 = [0.0, 0.0, 0.0] # initial position, [m]
 u0 = [1.0, 0.0, 0.0] # initial velocity, [m/s]
@@ -38,7 +49,7 @@ tspan_electron = (0.0, 15.0)
 param_proton = prepare(grid, E, B, species=Proton)
 tspan_proton = (0.0, 10.0)
 
-## Solve for the trajectories
+### Solve for the trajectories
 
 prob_e = ODEProblem(trace!, stateinit, tspan_electron, param_electron)
 prob_p = ODEProblem(trace!, stateinit, tspan_proton, param_proton)
@@ -46,9 +57,8 @@ prob_p = ODEProblem(trace!, stateinit, tspan_proton, param_proton)
 sol_e = solve(prob_e, Tsit5(); save_idxs=[1,2,3])
 sol_p = solve(prob_p, Tsit5(); save_idxs=[1,2,3])
 
-## Visualization
+### Visualization
 
-xyz = plot(sol_e, vars=(1,2,3), xlabel="x", ylabel="y", zlabel="z",
-   aspect_ratio=:equal, label="electron", lw=2)
-
-plot!(sol_p, vars=(1,2,3), label="proton", lw=2, camera = (70,50))
+plot(sol_e, color=:tomato, label="electron")
+plot!(sol_p, color=:deepskyblue3, label="proton")
+axislegend()
