+++ docs/examples/basics/demo_multiple.jl
@@ -1,14 +1,22 @@
-# Tracing multiple charged particles in a static EM field.
-#
-# Hongyang Zhou, hyzhou@umich.edu
+# ---
+# title: Multiple particles
+# id: demo_multiple
+# date: 2023-04-20
+# author: "[Hongyang Zhou](https://github.com/henry2004y)"
+# julia: 1.9.0
+# description: Tracing multiple charged particles in a static EM field
+# ---
+
+using JSServe: Page # hide
+Page(exportable=true, offline=true) # hide
 
 using TestParticle
 using OrdinaryDiffEq
-using GLMakie
 using TestParticleMakie
+using WGLMakie
 using Random
 
-# For reproducible results
+## For reproducible results
 Random.seed!(1234)
 
 struct Trajectory{T<:AbstractFloat}
@@ -18,19 +26,19 @@ struct Trajectory{T<:AbstractFloat}
 end
 
 function trace(x, y, z, E, B; trajectories::Int=10)
-   # Initialize particles
+   ## Initialize particles
    x0 = [0.0, 0.0, 0.0] # initial position, [m]
    u0 = [1.0, 0.0, 0.0] # initial velocity, [m/s]
    stateinit = [x0..., u0...]
-   
+
    param = prepare(x, y, z, E, B, species=Electron)
    tspan = (0.0, 15.0)
 
    prob = ODEProblem(trace!, stateinit, tspan, param)
-   # Trajectory container
+   ## Trajectory container
    sols = Vector{Trajectory}(undef, trajectories)
-   # Sample from a Maxwellian with bulk speed 0 and thermal speed 1.0
-   vdf = Maxwellian(0.0, 1.0)
+   ## Sample from a Maxwellian with bulk speed 0 and thermal speed 1.0
+   vdf = Maxwellian([0.0, 0.0, 0.0], 1.0)
    v = sample(vdf, trajectories)
 
    for i in 1:trajectories
@@ -46,7 +54,7 @@ function trace(x, y, z, E, B; trajectories::Int=10)
    return sols
 end
 
-## Initialize grid and field
+### Initialize grid and field
 
 x = range(-10, 10, length=15)
 y = range(-10, 10, length=20)
@@ -59,23 +67,23 @@ E[3,:,:,:] .= 5e-13
 
 trajectories = 10
 
-## Solve for the trajectories
+### Solve for the trajectories
 
 sols = trace(x, y, z, E, B; trajectories)
 
-## Visualization
+### Visualization
 
 f = Figure(fontsize = 18)
 ax = Axis3(f[1, 1],
-   title = "Particle trajectories, EGI, t=1298s",
-   #xlabel = "X",
-   #ylabel = "Y",
-   #zlabel = "Z",
+   title = "Particle trajectories",
+   xlabel = "X",
+   ylabel = "Y",
+   zlabel = "Z",
    aspect = :data,
 )
 
 for i in eachindex(sols)
-   lines!(sols[i].x, sols[i].y, sols[i].z, label="$i")
+   lines!(ax, sols[i].x, sols[i].y, sols[i].z, label="$i")
 end
 
 f
