+++ docs/examples/config.json
@@ -1,7 +1,7 @@
 {
    "theme": "grid",
    "order": [
-     "single_particle"
+     "basics"
    ],
    "properties": {
      "notebook": "false"
